+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'christophe.arguel@free.fr'
 license          'All rights reserved'
 description      'Installs/Configures Openldap (>=2.4)'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'
+version          '0.2.0'
 
 supports 'redhat', '>= 6.0'
 supports 'centos', '>= 6.0'