+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of ca_openldap.
 
+## 0.2.0:
+
+* Add support for TLS certificates and the related key files deployed by the _certificate_ cookbook (see attr
+ibute _use_existing_certs_and_key_)
+
 ## 0.1.0:
 
 * Initial release of ca_openldap